Create SidewalkClass.py
An objected-oriented script that utilizes classes and objects to iterate through conditional statements and output the compliance status based on the input parameters. This script was created for testing purposes and is not to be used in ArcGIS Pro. The SidewalkCalculateField.py file in the SidewalkAccessibility Repository displays the code and format for the following script to be used in the Calculate Field tool for ArcGIS Pro.